requirements.txt
alembic==1.13.3
anthropic==0.37.1
blinker==1.8.2
email_validator==2.2.0
langchain==0.1.14
langchain-community==0.0.32
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
gunicorn
Jinja2==3.1.4
openai==1.53.0
psutil==5.9.8
python-dotenv==1.0.1
SQLAlchemy==2.0.36
WTForms==3.2.1
